# Vectorize `find_first_not_of`/`find_last_not_of` member functions (multiple characters overloads)