--- a/benchmarks/src/find_first_of.cpp
+++ b/benchmarks/src/find_first_of.cpp
@@ -17,9 +17,15 @@
 
 using namespace std;
 
-enum class AlgType { std_func, str_member_first, str_member_last };
+enum class AlgType {
+    std_func,
+    str_member_first,
+    str_member_last,
+    str_member_first_not,
+    str_member_last_not,
+};
 
-template <AlgType Alg, class T, T Start = T{'!'}>
+template <AlgType Alg, class T, T NeedleFillerBase = T{'a'}>
 void bm(benchmark::State& state) {
     const size_t Pos   = static_cast<size_t>(state.range(0));
     const size_t NSize = static_cast<size_t>(state.range(1));
@@ -29,24 +35,37 @@ void bm(benchmark::State& state) {
     using container = conditional_t<Alg == AlgType::std_func, vector<T, not_highly_aligned_allocator<T>>,
         basic_string<T, char_traits<T>, not_highly_aligned_allocator<T>>>;
 
-    constexpr T HaystackFiller{' '};
-    static_assert(HaystackFiller < Start, "The following iota() should not produce the haystack filler.");
+    constexpr size_t IncremantCap = 16;
 
-    container h(HSize, HaystackFiller);
+    constexpr T HaystackFillerBase = {' '};
+    static_assert(
+        NeedleFillerBase + IncremantCap <= HaystackFillerBase || HaystackFillerBase + IncremantCap <= NeedleFillerBase,
+        "Would match where shouldn't");
+
+    container h(HSize, T{0});
     container n(NSize, T{0});
 
-    if (NSize - 1 > static_cast<size_t>(numeric_limits<T>::max()) - static_cast<size_t>(Start)) {
-        puts("ERROR: The following iota() would overflow.");
-        abort();
+    for (size_t i = 0, m = n.size(); i != m; ++i) {
+        n[i] = NeedleFillerBase + i % IncremantCap;
     }
-
-    iota(n.begin(), n.end(), Start);
 
     if (Pos >= HSize || Which >= NSize) {
         abort();
     }
 
-    h[Pos] = n[Which];
+    if constexpr (Alg == AlgType::str_member_first_not || Alg == AlgType::str_member_last_not) {
+        for (size_t i = 0, m = h.size(); i != m; ++i) {
+            h[i] = NeedleFillerBase + (i + Which) % IncremantCap;
+        }
+
+        h[Pos] = HaystackFillerBase;
+    } else {
+        for (size_t i = 0, m = h.size(); i != m; ++i) {
+            h[i] = HaystackFillerBase + i % IncremantCap;
+        }
+
+        h[Pos] = n[Which];
+    }
 
     for (auto _ : state) {
         benchmark::DoNotOptimize(h);
@@ -55,6 +74,10 @@ void bm(benchmark::State& state) {
             benchmark::DoNotOptimize(h.find_first_of(n));
         } else if constexpr (Alg == AlgType::str_member_last) {
             benchmark::DoNotOptimize(h.find_last_of(n));
+        } else if constexpr (Alg == AlgType::str_member_first_not) {
+            benchmark::DoNotOptimize(h.find_first_not_of(n));
+        } else if constexpr (Alg == AlgType::str_member_last_not) {
+            benchmark::DoNotOptimize(h.find_last_not_of(n));
         } else {
             benchmark::DoNotOptimize(find_first_of(h.begin(), h.end(), n.begin(), n.end()));
         }
@@ -81,5 +104,13 @@ BENCHMARK(bm<AlgType::str_member_first, char32_t, U'\x03B1'>)->Apply(common_args
 BENCHMARK(bm<AlgType::str_member_last, char>)->Apply(common_args);
 BENCHMARK(bm<AlgType::str_member_last, wchar_t>)->Apply(common_args);
 BENCHMARK(bm<AlgType::str_member_last, wchar_t, L'\x03B1'>)->Apply(common_args);
+
+BENCHMARK(bm<AlgType::str_member_first_not, char>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_first_not, wchar_t>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_first_not, wchar_t, L'\x03B1'>)->Apply(common_args);
+
+BENCHMARK(bm<AlgType::str_member_last_not, char>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_last_not, wchar_t>)->Apply(common_args);
+BENCHMARK(bm<AlgType::str_member_last_not, wchar_t, L'\x03B1'>)->Apply(common_args);
 
 BENCHMARK_MAIN();