--- a/stl/src/vector_algorithms.cpp
+++ b/stl/src/vector_algorithms.cpp
@@ -3552,6 +3552,7 @@ namespace {
 
                         if (_Bingo != 0) {
                             unsigned long _Tmp;
+                            // CodeQL [SM02313] _Tmp is always initialized: we just tested `if (_Bingo != 0)`.
                             _BitScanForward(&_Tmp, _Bingo);
                             if (_Found_pos > static_cast<int>(_Tmp)) {
                                 _Found_pos = _Tmp;
@@ -4083,6 +4084,7 @@ namespace {
 
                         if (_Bingo != 0) {
                             unsigned long _Tmp;
+                            // CodeQL [SM02313] _Tmp is always initialized: we just tested `if (_Bingo != 0)`.
                             _BitScanReverse(&_Tmp, _Bingo);
                             _Found_pos = _Tmp;
                         }