--- a/tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
+++ b/tests/std/tests/VSO_0000000_vector_algorithms/test.cpp
@@ -1062,12 +1062,25 @@ void test_bitset(mt19937_64& gen) {
 }
 
 template <class T>
-void test_case_string_find_first_of(const basic_string<T>& input_haystack, const basic_string<T>& input_needle) {
-    auto expected_iter = last_known_good_find_first_of(
-        input_haystack.begin(), input_haystack.end(), input_needle.begin(), input_needle.end());
-    auto expected = (expected_iter != input_haystack.end()) ? expected_iter - input_haystack.begin() : ptrdiff_t{-1};
-    auto actual   = static_cast<ptrdiff_t>(input_haystack.find_first_of(input_needle));
-    assert(expected == actual);
+size_t last_known_good_find_first_of(const basic_string<T>& h, const basic_string<T>& n) {
+    for (size_t pos = 0, pos_max = h.size(); pos != pos_max; ++pos) {
+        if (n.find(h[pos]) != basic_string<T>::npos) {
+            return pos;
+        }
+    }
+
+    return basic_string<T>::npos;
+}
+
+template <class T>
+size_t last_known_good_find_first_not_of(const basic_string<T>& h, const basic_string<T>& n) {
+    for (size_t pos = 0, pos_max = h.size(); pos != pos_max; ++pos) {
+        if (n.find(h[pos]) == basic_string<T>::npos) {
+            return pos;
+        }
+    }
+
+    return basic_string<T>::npos;
 }
 
 template <class T>
@@ -1084,9 +1097,43 @@ size_t last_known_good_find_last_of(const basic_string<T>& h, const basic_string
 }
 
 template <class T>
+size_t last_known_good_find_last_not_of(const basic_string<T>& h, const basic_string<T>& n) {
+    size_t pos = h.size();
+    while (pos != 0) {
+        --pos;
+        if (n.find(h[pos]) == basic_string<T>::npos) {
+            return pos;
+        }
+    }
+
+    return basic_string<T>::npos;
+}
+
+template <class T>
+void test_case_string_find_first_of(const basic_string<T>& input_haystack, const basic_string<T>& input_needle) {
+    size_t expected = last_known_good_find_first_of(input_haystack, input_needle);
+    size_t actual   = input_haystack.find_first_of(input_needle);
+    assert(expected == actual);
+}
+
+template <class T>
+void test_case_string_find_first_not_of(const basic_string<T>& input_haystack, const basic_string<T>& input_needle) {
+    size_t expected = last_known_good_find_first_not_of(input_haystack, input_needle);
+    size_t actual   = input_haystack.find_first_not_of(input_needle);
+    assert(expected == actual);
+}
+
+template <class T>
 void test_case_string_find_last_of(const basic_string<T>& input_haystack, const basic_string<T>& input_needle) {
     size_t expected = last_known_good_find_last_of(input_haystack, input_needle);
     size_t actual   = input_haystack.find_last_of(input_needle);
+    assert(expected == actual);
+}
+
+template <class T>
+void test_case_string_find_last_not_of(const basic_string<T>& input_haystack, const basic_string<T>& input_needle) {
+    size_t expected = last_known_good_find_last_not_of(input_haystack, input_needle);
+    size_t actual   = input_haystack.find_last_not_of(input_needle);
     assert(expected == actual);
 }
 
@@ -1178,6 +1225,8 @@ void test_basic_string_dis(mt19937_64& gen, D& dis) {
 
         test_case_string_find_first_of(input_haystack, input_needle);
         test_case_string_find_last_of(input_haystack, input_needle);
+        test_case_string_find_first_not_of(input_haystack, input_needle);
+        test_case_string_find_last_not_of(input_haystack, input_needle);
         test_case_string_find_str(input_haystack, input_needle);
         test_case_string_rfind_str(input_haystack, input_needle);
 
@@ -1185,6 +1234,8 @@ void test_basic_string_dis(mt19937_64& gen, D& dis) {
             input_needle.push_back(static_cast<T>(dis(gen)));
             test_case_string_find_first_of(input_haystack, input_needle);
             test_case_string_find_last_of(input_haystack, input_needle);
+            test_case_string_find_first_not_of(input_haystack, input_needle);
+            test_case_string_find_last_not_of(input_haystack, input_needle);
             test_case_string_find_str(input_haystack, input_needle);
             test_case_string_rfind_str(input_haystack, input_needle);