@AlexGuteniev 2025-02-20
> @AlexGuteniev Just checking, are you planning to make changes here in the near future?

I'm not sure. I guess I should close the PR if I won't find motivation to look into it next weekend.