@StephanTLavavej 2025-03-07
Thanks! :green_heart: I pushed minor changes, the most significant being CodeQL warning suppressions to avoid a headache in a few months.

5950X speedups look great:

Benchmark                                | Before  |   After | Speedup
-----------------------------------------|---------|---------|--------
`first_not, char>/2/3`                   | 9.60 ns | 9.68 ns | 0.99
`first_not, char>/6/81`                  | 27.8 ns | 21.9 ns | 1.27
`first_not, char>/7/4`                   | 11.1 ns | 18.3 ns | 0.61
`first_not, char>/9/3`                   | 11.7 ns | 16.3 ns | 0.72
`first_not, char>/22/5`                  | 19.2 ns | 16.7 ns | 1.15
`first_not, char>/58/2`                  | 33.1 ns | 17.3 ns | 1.91
`first_not, char>/75/85`                 | 65.3 ns | 45.0 ns | 1.45
`first_not, char>/102/4`                 | 57.7 ns | 18.1 ns | 3.19
`first_not, char>/200/46`                |  109 ns | 43.3 ns | 2.52
`first_not, char>/325/1`                 |  153 ns | 26.7 ns | 5.73
`first_not, char>/400/50`                |  197 ns | 57.1 ns | 3.45
`first_not, char>/1011/11`               |  453 ns | 93.1 ns | 4.87
`first_not, char>/1280/46`               |  575 ns |  120 ns | 4.79
`first_not, char>/1502/23`               |  661 ns |  129 ns | 5.12
`first_not, char>/2203/54`               |  963 ns |  175 ns | 5.50
`first_not, char>/3056/7`                | 1351 ns |  209 ns | 6.46
`first_not, wchar_t>/2/3`                | 8.96 ns | 8.53 ns | 1.05
`first_not, wchar_t>/6/81`               | 29.5 ns | 52.5 ns | 0.56
`first_not, wchar_t>/7/4`                | 11.4 ns | 15.4 ns | 0.74
`first_not, wchar_t>/9/3`                | 12.0 ns | 15.7 ns | 0.76
`first_not, wchar_t>/22/5`               | 18.7 ns | 15.9 ns | 1.18
`first_not, wchar_t>/58/2`               | 33.3 ns | 17.6 ns | 1.89
`first_not, wchar_t>/75/85`              | 65.4 ns | 45.8 ns | 1.43
`first_not, wchar_t>/102/4`              | 58.1 ns | 23.6 ns | 2.46
`first_not, wchar_t>/200/46`             |  109 ns | 52.0 ns | 2.10
`first_not, wchar_t>/325/1`              |  153 ns | 39.6 ns | 3.86
`first_not, wchar_t>/400/50`             |  197 ns | 66.3 ns | 2.97
`first_not, wchar_t>/1011/11`            |  449 ns |  118 ns | 3.81
`first_not, wchar_t>/1280/46`            |  571 ns |  153 ns | 3.73
`first_not, wchar_t>/1502/23`            |  666 ns |  164 ns | 4.06
`first_not, wchar_t>/2203/54`            |  967 ns |  229 ns | 4.22
`first_not, wchar_t>/3056/7`             | 1325 ns |  290 ns | 4.57
`first_not, wchar_t, L'\x03B1'>/2/3`     | 9.64 ns | 9.88 ns | 0.98
`first_not, wchar_t, L'\x03B1'>/6/81`    | 37.3 ns | 21.3 ns | 1.75
`first_not, wchar_t, L'\x03B1'>/7/4`     | 15.3 ns | 15.3 ns | 1.00
`first_not, wchar_t, L'\x03B1'>/9/3`     | 16.6 ns | 15.7 ns | 1.06
`first_not, wchar_t, L'\x03B1'>/22/5`    | 34.8 ns | 15.8 ns | 2.20
`first_not, wchar_t, L'\x03B1'>/58/2`    | 55.8 ns | 17.5 ns | 3.19
`first_not, wchar_t, L'\x03B1'>/75/85`   |  247 ns | 88.1 ns | 2.80
`first_not, wchar_t, L'\x03B1'>/102/4`   |  122 ns | 22.0 ns | 5.55
`first_not, wchar_t, L'\x03B1'>/200/46`  |  593 ns |  136 ns | 4.36
`first_not, wchar_t, L'\x03B1'>/325/1`   |  223 ns | 39.9 ns | 5.59
`first_not, wchar_t, L'\x03B1'>/400/50`  | 1155 ns |  277 ns | 4.17
`first_not, wchar_t, L'\x03B1'>/1011/11` | 2129 ns |  236 ns | 9.02
`first_not, wchar_t, L'\x03B1'>/1280/46` | 3510 ns |  706 ns | 4.97
`first_not, wchar_t, L'\x03B1'>/1502/23` | 3724 ns |  456 ns | 8.17
`first_not, wchar_t, L'\x03B1'>/2203/54` | 6200 ns | 1384 ns | 4.48
`first_not, wchar_t, L'\x03B1'>/3056/7`  | 4484 ns |  268 ns | 16.73
`last_not, char>/2/3`                    | 8.11 ns | 8.30 ns | 0.98
`last_not, char>/6/81`                   | 26.6 ns | 21.6 ns | 1.23
`last_not, char>/7/4`                    | 9.64 ns | 16.6 ns | 0.58
`last_not, char>/9/3`                    | 9.88 ns | 14.3 ns | 0.69
`last_not, char>/22/5`                   | 13.8 ns | 15.0 ns | 0.92
`last_not, char>/58/2`                   | 21.3 ns | 14.9 ns | 1.43
`last_not, char>/75/85`                  | 47.0 ns | 40.2 ns | 1.17
`last_not, char>/102/4`                  | 32.0 ns | 16.6 ns | 1.93
`last_not, char>/200/46`                 | 68.6 ns | 40.5 ns | 1.69
`last_not, char>/325/1`                  | 96.5 ns | 27.4 ns | 3.52
`last_not, char>/400/50`                 |  113 ns | 54.7 ns | 2.07
`last_not, char>/1011/11`                |  234 ns | 90.8 ns | 2.58
`last_not, char>/1280/46`                |  297 ns |  117 ns | 2.54
`last_not, char>/1502/23`                |  344 ns |  124 ns | 2.77
`last_not, char>/2203/54`                |  498 ns |  176 ns | 2.83
`last_not, char>/3056/7`                 |  675 ns |  197 ns | 3.43
`last_not, wchar_t>/2/3`                 | 8.98 ns | 8.78 ns | 1.02
`last_not, wchar_t>/6/81`                | 47.2 ns | 33.3 ns | 1.42
`last_not, wchar_t>/7/4`                 | 11.6 ns | 14.0 ns | 0.83
`last_not, wchar_t>/9/3`                 | 10.9 ns | 13.7 ns | 0.80
`last_not, wchar_t>/22/5`                | 15.1 ns | 14.4 ns | 1.05
`last_not, wchar_t>/58/2`                | 25.4 ns | 17.5 ns | 1.45
`last_not, wchar_t>/75/85`               | 65.2 ns | 44.5 ns | 1.47
`last_not, wchar_t>/102/4`               | 40.1 ns | 22.5 ns | 1.78
`last_not, wchar_t>/200/46`              | 98.3 ns | 50.0 ns | 1.97
`last_not, wchar_t>/325/1`               |  120 ns | 38.5 ns | 3.12
`last_not, wchar_t>/400/50`              |  163 ns | 64.2 ns | 2.54
`last_not, wchar_t>/1011/11`             |  360 ns |  114 ns | 3.16
`last_not, wchar_t>/1280/46`             |  443 ns |  148 ns | 2.99
`last_not, wchar_t>/1502/23`             |  512 ns |  158 ns | 3.24
`last_not, wchar_t>/2203/54`             |  752 ns |  221 ns | 3.40
`last_not, wchar_t>/3056/7`              |  999 ns |  270 ns | 3.70
`last_not, wchar_t, L'\x03B1'>/2/3`      | 8.98 ns | 8.94 ns | 1.00
`last_not, wchar_t, L'\x03B1'>/6/81`     | 45.7 ns | 20.5 ns | 2.23
`last_not, wchar_t, L'\x03B1'>/7/4`      | 15.3 ns | 13.7 ns | 1.12
`last_not, wchar_t, L'\x03B1'>/9/3`      | 17.0 ns | 13.9 ns | 1.22
`last_not, wchar_t, L'\x03B1'>/22/5`     | 37.2 ns | 14.1 ns | 2.64
`last_not, wchar_t, L'\x03B1'>/58/2`     | 63.4 ns | 17.3 ns | 3.66
`last_not, wchar_t, L'\x03B1'>/75/85`    |  266 ns | 94.7 ns | 2.81
`last_not, wchar_t, L'\x03B1'>/102/4`    |  133 ns | 21.7 ns | 6.13
`last_not, wchar_t, L'\x03B1'>/200/46`   |  606 ns |  148 ns | 4.09
`last_not, wchar_t, L'\x03B1'>/325/1`    |  292 ns | 43.2 ns | 6.76
`last_not, wchar_t, L'\x03B1'>/400/50`   | 1199 ns |  312 ns | 3.84
`last_not, wchar_t, L'\x03B1'>/1011/11`  | 2190 ns |  342 ns | 6.40
`last_not, wchar_t, L'\x03B1'>/1280/46`  | 3783 ns |  829 ns | 4.56
`last_not, wchar_t, L'\x03B1'>/1502/23`  | 3818 ns |  614 ns | 6.22
`last_not, wchar_t, L'\x03B1'>/2203/54`  | 6337 ns | 1593 ns | 3.98
`last_not, wchar_t, L'\x03B1'>/3056/7`   | 5124 ns |  269 ns | 19.05